Add initial program skeleton
This app is to be used for gathering bibliographic data from pictures of
the title page and colophon of books and other written material.

At the moment it contains a starting activity with input for the book id
and buttons to start camera activity for taking pictures.

It lacks the functionality of sending the captured images of for
analysis and receiving and displaying the analysis results.