Upload images with the send button
From android 9 all requests for non-https addresses are blocked by
default. This behaviour can be configured via a network security
configuration file which can have both a blacklist and a whitelist.
It is recommended to keep a list of domains for which non-secure traffic
is permitted rather than the other way around.
I just need to find to find a way of getting the build config variable
from gradle into the xml file.

https://developer.android.com/training/articles/security-config